Move largest directories logic into separate sections file.
